adds abstracted newPromptHandler, modifies some of the functions in storageHandler and databaseHandler, replaces useContext with useState on generate page, simplifies code on generate page to only call newPromptHandler
